Color Curve
Added a linear color curve used to change the baby's material based on its tiredness level. I changed the Child_BP in order to implement the dynamic material, but I pushed a copy of the blueprint to avoid merge issues. My changes will need to be copied over to the real blueprint later.